+++ user/token.go
@@ -4,9 +4,11 @@ import (
 	"errors"
 	"net/http"
 	"strconv"
+	"strings"
 	"time"
 
 	jwt "github.com/dgrijalva/jwt-go"
+	"github.com/google/uuid"
 )
 
 type (
@@ -24,7 +26,11 @@ type (
 	TokenData struct {
 		IsServer     bool   `json:"isserver"`
 		UserId       string `json:"userid"`
+		Email        string `json:"email"`
+		Name         string `json:"name"`
+		IsClinic     bool   `json:"isclinic"`
 		DurationSecs int64  `json:"-"`
+		Audience     string `json:"audience"`
 	}
 
 	TokenConfig struct {
@@ -67,9 +73,28 @@ func CreateSessionToken(data *TokenData, config TokenConfig) (*SessionToken, err
 	} else {
 		claims["svr"] = "no"
 	}
+	// Add claims specific to our 3rd party services
+	if strings.ToUpper(data.Audience) == "ZENDESK" {
+		if data.IsClinic {
+			claims["organization"] = "Psad"
+		} else {
+			claims["organization"] = "Patient"
+			claims["tags"] = "patient"
+		}
+		claims["aud"] = "zendesk"
+	}
 	claims["usr"] = data.UserId
+	if data.Name != "" {
+		claims["name"] = data.Name
+	}
+	if data.Email != "" {
+		claims["email"] = data.Email
+	}
+
 	claims["dur"] = data.DurationSecs
 	claims["exp"] = expiresAt
+	claims["iat"] = createdAt
+	claims["jti"] = uuid.New()
 
 	tokenString, err := token.SignedString([]byte(config.Secret))
 	if err != nil {
@@ -128,10 +153,21 @@ func UnpackSessionTokenAndVerify(id string, secret string) (*TokenData, error) {
 	}
 	userId := claims["usr"].(string)
 
+	email, ok := claims["email"].(string)
+	if !ok {
+		email = ""
+	}
+	name, ok := claims["name"].(string)
+	if !ok {
+		name = email
+	}
+
 	return &TokenData{
 		IsServer:     isServer,
 		DurationSecs: durationSecs,
 		UserId:       userId,
+		Email:        email,
+		Name:         name,
 	}, nil
 }